requirements.txt
altair==5.3.0
annotated-types==0.7.0
blinker==1.8.2
cachetools==5.3.3
click==8.1.7
distro==1.9.0
exa-py==1.0.12
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
httplib2==0.22.0
numpy==1.26.4
openai==1.30.3
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pyparsing==3.1.2
python-dotenv==1.0.1
pytz==2024.1
rsa==4.9
smmap==5.0.1
streamlit==1.35.0
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tqdm==4.66.4
tzdata==2024.1
uritemplate==4.1.1
watchdog==4.0.1